launch animation working! angle of aim off, not sure how to fix
